refactor: extract kbm (teaching-journal notes) out of courses.py
Sixth phase — the first slice carved out of the courses.py god-blueprint,
per the decomposition plan. KbmNote/KbmActivityType were previously
defined inside app/models/course.py (a hub file with 6 unrelated
concerns); they're now app/kbm/models.py, a genuinely standalone
Django-style app with no dependency on the rest of courses.py beyond a
plain Course lookup.

- Split KbmNote/KbmActivityType out of app/models/course.py (which
  keeps Course/AcademicYear/Link/File/UserCourseOrder/enrollments for
  now — those move in later phases) into app/kbm/models.py
- Moved the 4 KBM-notes routes (get/create/update/delete), previously
  a comment-delimited block in app/blueprints/courses.py, into
  app/kbm/routes.py as their own blueprint. URL paths are byte-for-byte
  identical (/api/courses/<id>/kbm-notes, /api/kbm-notes/<id>) since
  these are called from JS via hardcoded fetch URLs, not url_for() —
  only the internal Flask endpoint name changes (courses.* -> kbm.*),
  confirmed safe by grepping for any url_for() reference to the old
  endpoint names (none exist)
- app/models/__init__.py sources KbmNote/KbmActivityType from
  app.kbm.models now; courses.py's aggregator import trimmed accordingly

Verified: create_app() builds with all 224 routes intact (4 moved from
courses.* to kbm.* endpoints, net count unchanged), SQLAlchemy metadata
still has 33 tables, and pytest is unchanged from baseline (25
pre-existing failures, 99 passing).
