+++ app/models/__init__.py
@@ -2,7 +2,8 @@ from app.core.extensions import db
 
 from .school import School, SchoolStatus
 from .user import User, UserRole
-from .course import Course, AcademicYear, Link, File, enrollments, UserCourseOrder, KbmNote, KbmActivityType
+from .course import Course, AcademicYear, Link, File, enrollments, UserCourseOrder
+from app.kbm.models import KbmNote, KbmActivityType
 from .content_folder import ContentFolder
 from app.quiz.models import (
     Quiz, Question, Option, QuizSubmission, Answer,
